back/requirements.txt
aiomysql==0.2.0
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
colorama==0.4.6
databases==0.7.0
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
pydantic==2.1.1
pydantic_core==2.4.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
alembic==1.11.2
cryptography==41.0.3
email-validator==2.0.0.post2
pyjwt==2.8.0
requests==2.31.0
pika==1.3.2
